+++ .gitignore
@@ -34,3 +34,4 @@
 /vendor/
 
 .vscode
+*-local.*
